[FEAT] Stomp 관련 헤더를 처리하는 Provider 클래스 추가, 헤더 처리 로직 재사용 위해 분리
